fix(terminal): 修复 DOM 终端缩放后宽度恢复
